// src/plugins/users/types.ts
/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/** A CouchDB `_users` user document (post-create includes hash fields). */
export interface UserDoc {
  _id: string;
  _rev?: string;
  name: string;
  type: 'user';
  roles: string[];
  [key: string]: unknown;
}

/** Form-collected fields used to build a create body. */
export interface UserDraft {
  name: string;
  password?: string;
  roles: string[];
}